+++ servente_generator/src/lib.rs
@@ -1,0 +1,7 @@
+// Copyright (C) 2023 Tristan Gerritsen <tristan@thewoosh.org>
+// All Rights Reserved.
+
+//! This crate can generate certain files, like transforming Markdown to HTML.
+
+#[cfg(feature = "markdown")]
+pub mod common_mark;